perf!(ServiceProviderRegistry): Bloom Schema (#308)
Reviewer @rvagg
Closes #307
This significantly reduces the size of ServiceProviderRegistry: 21,290
-> 17,751 (-3539)

#### Motivation
We want to allow all product attributes to be queriable on-chain.
Therefore we are removing the encoded productData. Mandatory schema will
be loosely enforced on-chain with a bloom filter.

#### Upgrade guide for synapse and curio
1. All `productInfo` previously abi-encoded is removed and everything is
now capability key-value store. Keys are `string`; values are `bytes`.
This affects `registerProvider`, `addProduct`, and `updateProduct`.
Unsigned integers should be encoded big-endian. Addresses should be
encoded as `bytes[20]`. Strings should be encoded utf-8. See the
examples in `test/PDPOffering.sol`. We don't validate these values
on-chain so if they don't look right, throw.
2. `getPDPOffering` is not how you get PDP product info anymore. Now use
`getAllProductCapabilities` which will return all keys and values for a
product.
3. `getActiveProvidersByProductType` and `getProvidersByProductType` now
also return capability values via `productCapabilityValues` in the
returned `ProviderWithProduct[]`.
4. `getProductCapabilities` no longer returns the `exists` `bool` array.
This array was misleading because it only indicated whether the bytes
were empty.
5. `getProductCapability` is removed. Use `productCapabilities` instead,
which only differs in not having `exists`.
6. `updatePDPServiceWithCapabilities` is removed. Use `updateProduct`.
7. Capability values cannot be empty. Exclude the key to signal that the
product does not have the capability.

#### Changes
* define and configure BloomSet with k=16
* move PDPOffering struct to a testing helper library
* redefine required schema as BloomSet16
* only enforce required schema probabilistically
* ensure synapse has a good method for fetching these keys all at once
* remove misleading `exists`
* remove ipni capability flags from required schema
* fix tests

---------

Co-authored-by: Rod Vagg <rod@vagg.org>
Co-authored-by: Jakub Sztandera <kubuxu@protonmail.ch>